create seperate components for user feed page
