+++ src/app/_components/userFeed/LeftSideBar.tsx
@@ -1,0 +1,187 @@
+import { Box, Typography, Button, Grid2 as Grid } from "@mui/material";
+import React from "react";
+import { STATIC_TEXTS } from "~/app/_components/static/staticText";
+import OwlSVG from "~/assets/svg/owl.svg";
+import WriteLogo from "~/assets/svg/writeLogo.svg";
+import MoreHorizOutlinedIcon from "@mui/icons-material/MoreHorizOutlined";
+import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
+import SearchIcon from "@mui/icons-material/Search";
+import MenuBookIcon from "@mui/icons-material/MenuBook";
+import StarOutlineIcon from "@mui/icons-material/StarOutline";
+import MessageIcon from "@mui/icons-material/Message";
+import BookmarkBorderOutlinedIcon from "@mui/icons-material/BookmarkBorderOutlined";
+import ShoppingCartOutlinedIcon from "@mui/icons-material/ShoppingCartOutlined";
+import SettingsOutlinedIcon from "@mui/icons-material/SettingsOutlined";
+import AccountCircleOutlinedIcon from "@mui/icons-material/AccountCircleOutlined";
+import userImage from "~/assets/images/user.jpeg";
+import { MENU_ITEMS } from "~/app/user-feed/_config/config";
+import Image from "next/image";
+
+const LeftSideBar = () => {
+  const ICONS_MAP = {
+    HomeIcon: HomeOutlinedIcon,
+    SearchIcon: SearchIcon,
+    MenuBookIcon: MenuBookIcon,
+    StarOutlineIcon: StarOutlineIcon,
+    MessagesIcon: MessageIcon,
+    BookmarkBorderOutlinedIcon: BookmarkBorderOutlinedIcon,
+    ShoppingCartOutlinedIcon: ShoppingCartOutlinedIcon,
+    SettingsOutlinedIcon: SettingsOutlinedIcon,
+    AccountCircleOutlinedIcon: AccountCircleOutlinedIcon,
+  };
+
+  return (
+    <Grid
+      columns={1}
+      sx={{
+        height: "100%",
+        width: "100%",
+        spacing: 3,
+        backgroundColor: "white",
+        position: "relative",
+        py: 1,
+        px: "8px",
+      }}
+    >
+      <Box
+        sx={{
+          display: "flex",
+          justifyContent: "start",
+          alignItems: "center",
+          px: 2,
+        }}
+      >
+        <OwlSVG />
+        <Typography variant="h5" fontWeight="bold" color="primary">
+          {STATIC_TEXTS.APP_TITLE}
+        </Typography>
+      </Box>
+      <Box sx={{ borderBottom: 1, borderColor: "grey.300" }}></Box>
+      <Box
+        sx={{
+          marginTop: 4,
+        }}
+      >
+        <Button
+          startIcon={<WriteLogo />}
+          variant="contained"
+          size="small"
+          fullWidth
+          sx={{
+            display: "flex",
+            height: "40px",
+            backgroundColor: "primary.main",
+            justifyContent: "start",
+            alignItems: "center",
+            px: "12px",
+          }}
+        >
+          <Typography
+            variant="h6"
+            color="#fff"
+            style={{
+              fontSize: "16px",
+            }}
+          >
+            {STATIC_TEXTS.USER_FEED.BUTTONS.WRITE}
+          </Typography>
+        </Button>
+      </Box>
+      <Box
+        sx={{
+          marginTop: 2,
+        }}
+      >
+        {MENU_ITEMS.map((item, index) => {
+          const IconComponent = ICONS_MAP[item.icon];
+
+          return (
+            <Box
+              key={index}
+              sx={{
+                display: "flex",
+                justifyContent: "start",
+                alignItems: "center",
+                px: "2px",
+                py: "4px",
+              }}
+            >
+              <Button
+                startIcon={<IconComponent />}
+                variant="text"
+                size="small"
+                fullWidth
+                sx={{
+                  display: "flex",
+                  height: "40px",
+                  justifyContent: "start",
+                  alignItems: "center",
+                  color: "text.secondary",
+                  marginLeft: "4px",
+                }}
+              >
+                <Typography
+                  variant="h6"
+                  sx={{
+                    fontSize: "18px",
+                    fontWeight: "500",
+                    color: "text.secondary",
+                  }}
+                >
+                  {item.label}
+                </Typography>
+              </Button>
+            </Box>
+          );
+        })}
+      </Box>
+      <Box
+        sx={{
+          display: "flex",
+          justifyContent: "space-between",
+          alignItems: "center",
+          width: "100%",
+          position: "absolute",
+          bottom: "0",
+        }}
+      >
+        <Box
+          sx={{
+            display: "flex",
+            justifyContent: "center",
+            alignItems: "center",
+          }}
+        >
+          <Image
+            alt="profile picture "
+            src={userImage}
+            width={50}
+            height={50}
+            style={{
+              borderRadius: "100%",
+            }}
+          ></Image>
+          <Typography
+            sx={{
+              fontSize: "17px",
+              fontWeight: "semibold",
+              color: "text.secondary",
+              marginLeft: "4px",
+            }}
+          >
+            Steve Jobs
+          </Typography>
+        </Box>
+        <Box>
+          <Button
+            startIcon={<MoreHorizOutlinedIcon />}
+            sx={{ color: "text.secondary" }}
+            size="small"
+          ></Button>
+        </Box>
+      </Box>
+    </Grid>
+  );
+};
+
+export default LeftSideBar;
