+++ src/theme.ts
@@ -1,6 +1,20 @@
 "use client";
 import { createTheme } from "@mui/material/styles";
 
+declare module "@mui/material/styles" {
+  interface PaletteOptions {
+    chip?: {
+      lightGray?: string;
+    };
+  }
+}
+
+declare module "@mui/material/styles" {
+  interface CommonColors {
+    lightGray?: string;
+  }
+}
+
 const theme = createTheme({
   palette: {
     mode: "light",
@@ -16,6 +30,7 @@ const theme = createTheme({
     common: {
       white: "#fff",
       black: "#101828",
+      lightGray: "#F2F3F5",
     },
     text: {
       primary: "#101828",
